1) Abstract The abstract was inacurate. We had this:
    This document discusses the importance of providing a formal semantics for ODRL.

I have now created this text:

    This document specifies the expected behaviour of an ODRL Evaluator, a piece of software that performs computations based on a set of policies and a certain state of the world. Both a monitoring scenario and an access control scenarios are considered.

2) Section 1 and 2.
I have not encountered contradiction in these sections with what we have discussed today. They can remain. They insist on the representation of the world with RDF --but we provide no example, though.

3a) Section 3, Consequences and Remedies
I have harmonized the terms Consequence and Remedy (boldface, capital) with the two previous concepts (Obligation, Condition). Same text, different format.
But I think we need examples with Consequences and Remedies.

3b) Section 3,
This text is anomalous (properties of what?). Constraints and Refinements were used without further introduction. I think we needed to list the basic concepts of the model. Therefore, I have replaced this text:

    Here follows a list of properties (with their possible values) needed by the ODRL Evaluator:

With a new one that introduces the missing objects. In addition, I had to make a reference to Constraints and Refinements --stressing what is the in Information Model.